(feat): update permissions

+++ app/Http/Livewire/Admin/Sites/Index.php
@@ -3,12 +3,13 @@
 namespace App\Http\Livewire\Admin\Sites;
 
 use App\Models\Site;
-
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Index extends Component
 {
+    use AuthorizesRequests;
     use WithPagination;
 
     public $search;
@@ -25,6 +26,8 @@ class Index extends Component
 
     public function destroy($id)
     {
+        $this->authorize('sites delete');
+
         $site = Site::findOrFail($id);
         $site->delete();
 
@@ -33,6 +36,8 @@ class Index extends Component
 
     public function render()
     {
+        $this->authorize('sites read');
+
         $sites = Site::orderBy('created_at', 'asc')->where('name', 'like', '%'.$this->search.'%')->paginate(2);
 
         return view('livewire.admin.sites.index', compact('sites'));
